fix(psci): typecast operands to match data type
This corrects the MISRA violation C2012-10.3:
The value of an expression shall not be assigned to an object with a
narrower essential type or of a different essential type category.
The condition is explicitly checked against 0U, appending 'U' and
typecasting for unsigned comparison.

State: not-upstreamable (fusa)
Link: https://review.trustedfirmware.org/c/TF-A/trusted-firmware-a/+/40460

+++ include/lib/el3_runtime/context_mgmt.h
@@ -26,7 +26,7 @@ void *cm_get_context_by_index(unsigned int cpu_idx,
 			      size_t security_state);
 void cm_set_context_by_index(unsigned int cpu_idx,
 			     void *context,
-			     unsigned int security_state);
+			     size_t security_state);
 void *cm_get_context(size_t security_state);
 void cm_set_context(void *context, uint32_t security_state);
 void cm_init_my_context(const struct entry_point_info *ep);
